+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - Update default ktlint version to 0.8.1
  - Fix extension version has no effect on used ktlint version
  - Add check task also depends on ktlintCheck task
+ - Add output report
+ - Add report type to extension
 
 ## [2.1.0] - 2017-07-5
 ### Added